requirements.txt
fastapi==0.128.0
httpx==0.28.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.12.5
pytest==8.4.2
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.47
uvicorn==0.40.0
